Refactor: move RefreshTokenProvider
